requirements.txt
streamlit==1.32.0
llama-cpp-python==0.2.65
Pillow==10.2.0
pypdfium2==4.27.0
numpy==1.26.4
surya-ocr
torch
torchvision
torchaudio